requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.2.0
attrs==22.1.0
Babel==2.11.0
backcall==0.2.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi
cffi==1.15.1
charset-normalizer==2.1.1
clip
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
flit_core
ftfy==6.1.1
idna==3.4
importlib-metadata==5.1.0
importlib-resources==5.10.0
ipykernel==6.17.1
ipython==8.7.0
ipython-genutils==0.2.0
jedi==0.18.2
Jinja2==3.1.2
json5==0.9.10
jsonschema==4.17.3
jupyter-server==1.23.3
jupyter_client==7.4.7
jupyter_core==5.1.0
jupyterlab==3.5.0
jupyterlab-pygments==0.2.2
jupyterlab_server==2.16.3
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.5
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy
packaging==21.3
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pkgutil_resolve_name==1.3.10
platformdirs==2.5.4
prometheus-client==0.15.0
prompt-toolkit==3.0.33
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
python-dateutil==2.8.2
pytz==2022.6
pyzmq==24.0.1
regex==2022.10.31
requests==2.28.1
Send2Trash==1.8.0
six
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
terminado==0.17.0
tinycss2==1.2.1
tomli==2.0.1
torch==1.7.1
torchvision==0.8.2
tornado==6.2
tqdm==4.64.1
traitlets==5.6.0
typing_extensions
urllib3==1.26.13
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.2
zipp==3.11.0
